chore: move get ui config hook to shared
