+++ packages/apps/human-app/frontend/src/modules/worker/jobs/hooks/index.ts
@@ -3,5 +3,4 @@ export * from './use-jobs-notifications';
 export * from './use-job-types-oracles-table';
 export * from './use-jobs-filter-store';
 export * from './use-my-jobs-filter-store';
-export * from './use-get-ui-config';
 export * from './use-get-my-jobs-data';
