+++ packages/apps/human-app/frontend/src/modules/worker/jobs/jobs.page.tsx
@@ -8,9 +8,9 @@ import { useIsMobile } from '@/shared/hooks/use-is-mobile';
 import { useColorMode } from '@/shared/contexts/color-mode';
 import { NoRecords } from '@/shared/components/ui/no-records';
 import { PageCardLoader } from '@/shared/components/ui/page-card';
+import { useGetUiConfig } from '@/shared/hooks';
 import { useGetOracles } from '../hooks';
 import { useGetOraclesNotifications } from '../hooks/use-get-oracles-notifications';
-import { useGetUiConfig } from './hooks';
 import { TabPanel } from './components';
 import { AvailableJobsView } from './available-jobs';
 import { MyJobsView } from './my-jobs/my-jobs-view';
